requirements.txt
fastapi
uvicorn 
google-api-python-client
google-auth-oauthlib
pytesseract
elasticsearch
PyPDF2
requests
pdfplumber